Integrate `nannou_core` into main `nannou` crate
This removes the color, math, rand and geom modules in favour of
depending on the `nannou_core` crate. The `nannou_core` modules are
re-exported within `nannou` in order to provide a near identical API to
the existing one.

The majority of changes here revolve around the change from `cgmath` to
`glam`. The most notable changes include:

- `geom::Vector2` -> `glam::Vec2`
- `cgmath::Matrix4<f32>` -> `glam::Mat4`
- `magnitude` -> `length`

This change also removes the generic `Scalar` type parameter from the
`Draw` API and related items in the hope of improving downstream compile
times.

Also removes the seemingly unused `geom::graph` module. Use-cases for
this module and its abstractions were mostly made reduntatnt upon
allowing for transformations to be applied to the `Draw` instance
itself.